check next authority id to determine new threshold in new session started end to end test for threshold
