+++ standalone/runtime/src/lib.rs
@@ -236,7 +236,8 @@ impl pallet_timestamp::Config for Runtime {
 	type WeightInfo = ();
 }
 
-#[cfg(feature = "integration-tests")]
+
+#[cfg(not(feature = "integration-tests"))]
 parameter_types! {
 	// How often we trigger a new session.
 	// during integration tests, we use manual sessions.
@@ -244,7 +245,7 @@ parameter_types! {
 	pub const Offset: BlockNumber = 0;
 }
 
-#[cfg(not(feature = "integration-tests"))]
+#[cfg(feature = "integration-tests")]
 parameter_types! {
 	// How often we trigger a new session.
 	pub const Period: BlockNumber = 3 * MINUTES;
